adjustments in some variables

+++ src/main/java/com/saulProgramado/NunesSportsAPI/domain/Product/ProductConsultData.java
@@ -1,6 +1,8 @@
 package com.saulProgramado.NunesSportsAPI.domain.Product;
 
-public record ProductConsultData(Long id, String name, String code, String describe, Double price) {
+import java.math.BigDecimal;
+
+public record ProductConsultData(Long id, String name, String code, String describe, BigDecimal price) {
 	
 	public ProductConsultData(Product product) {
 		this(product.getId(), product.getName(), product.getCode(), product.getDescribe(), product.getPrice());
