+++ src/main/java/com/saulProgramado/NunesSportsAPI/domain/Product/ProductDetailsData.java
@@ -1,6 +1,8 @@
 package com.saulProgramado.NunesSportsAPI.domain.Product;
 
-public record ProductDetailsData(Long id, String nome, String code, String describe, Double price) {
+import java.math.BigDecimal;
+
+public record ProductDetailsData(Long id, String nome, String code, String describe, BigDecimal price) {
 
 	public ProductDetailsData(Product product) {
 		this(product.getId(), product.getName(), product.getCode(), product.getDescribe(), product.getPrice());
